requirements.txt
librosa>=0.10.0
numpy>=1.24.0
PyQt5>=5.15.0
python-vlc>=3.0.0
